+++ allauth/account/tests.py
@@ -11,7 +11,10 @@ from django.core.urlresolvers import reverse
 from django.test.client import Client
 from django.core import mail
 from django.test.client import RequestFactory
-from django.contrib.auth.models import AnonymousUser
+from django.contrib.auth.models import AnonymousUser, AbstractUser
+from django.db import models
+
+import unittest
 
 from allauth.account.forms import BaseSignupForm
 from allauth.account.models import EmailAddress, EmailConfirmation
@@ -21,6 +24,10 @@ from . import app_settings
 
 from .auth_backends import AuthenticationBackend
 from .adapter import get_adapter
+from .utils import url_str_to_user_pk, user_pk_to_url_str
+
+import uuid
+import mock
 
 
 @override_settings(
@@ -199,7 +206,8 @@ class AccountTests(TestCase):
         resp = self.client.post(url,
                                 {'password1': 'newpass123',
                                  'password2': 'newpass123'})
-        self.assertRedirects(resp, reverse('account_reset_password_from_key_done'))
+        self.assertRedirects(resp,
+                             reverse('account_reset_password_from_key_done'))
 
         # Check the new password is in effect
         user = get_user_model().objects.get(pk=user.pk)
@@ -215,7 +223,8 @@ class AccountTests(TestCase):
 
         # Same should happen when accessing the page directly
         response = self.client.get(url)
-        self.assertTemplateUsed(response, 'account/password_reset_from_key.html')
+        self.assertTemplateUsed(response,
+                                'account/password_reset_from_key.html')
         self.assertTrue(response.context_data['token_fail'])
 
         # When in XHR views, it should respond with a 400 bad request
@@ -327,8 +336,9 @@ class AccountTests(TestCase):
                                 {'login': 'john',
                                  'password': 'doe'})
         self.assertRedirects(resp, reverse('account_email_verification_sent'))
-        self.assertEqual(resp['location'],
-                         'http://testserver' + reverse('account_email_verification_sent'))
+        self.assertEqual(
+            resp['location'],
+            'http://testserver' + reverse('account_email_verification_sent'))
 
     def test_login_inactive_account(self):
         """
@@ -338,7 +348,8 @@ class AccountTests(TestCase):
         regardless of their verified state.
         """
         # Inactive and verified user account
-        user = get_user_model().objects.create(username='john', is_active=False)
+        user = get_user_model().objects.create(username='john',
+                                               is_active=False)
         user.set_password('doe')
         user.save()
         EmailAddress.objects.create(user=user,
@@ -610,7 +621,6 @@ class BaseSignupFormTests(TestCase):
         self.assertTrue(form.is_valid())
 
 
-
 class AuthenticationBackendTests(TestCase):
 
     def setUp(self):
@@ -669,3 +679,35 @@ class AuthenticationBackendTests(TestCase):
                 username=user.username,
                 password=user.username).pk,
             user.pk)
+
+
+class UtilsTests(TestCase):
+    def setUp(self):
+        if hasattr(models, 'UUIDField'):
+            self.user_id = uuid.uuid4().hex
+
+            class UUIDUser(AbstractUser):
+                id = models.UUIDField(primary_key=True,
+                                      default=uuid.uuid4,
+                                      editable=False)
+
+                class Meta(AbstractUser.Meta):
+                    swappable = 'AUTH_USER_MODEL'
+        else:
+            UUIDUser = get_user_model()
+        self.UUIDUser = UUIDUser
+
+    @unittest.skipUnless(hasattr(models, 'UUIDField'),
+                         reason="No UUIDField in this django version")
+    def test_url_str_to_pk_identifies_UUID_as_stringlike(self):
+        with mock.patch('allauth.account.utils.get_user_model') as mocked_gum:
+            mocked_gum.return_value = self.UUIDUser
+            self.assertEqual(url_str_to_user_pk(self.user_id),
+                             self.user_id)
+
+    def test_pk_to_url_string_identifies_UUID_as_stringlike(self):
+        user = self.UUIDUser(
+            is_active=True,
+            email='john@doe.com',
+            username='john')
+        self.assertEquals(user_pk_to_url_str(user), str(user.pk))