+++ allauth/socialaccount/tests.py
@@ -17,11 +17,12 @@ from django.contrib.auth.models import AnonymousUser
 from ..tests import MockedResponse, mocked_response
 from ..account import app_settings as account_settings
 from ..account.models import EmailAddress
-from ..account.utils import user_email
+from ..account.utils import user_email, user_username
 from ..utils import get_user_model, get_current_site
 
 from .models import SocialApp, SocialAccount, SocialLogin
 from .helpers import complete_social_login
+from .views import signup
 
 
 def create_oauth_tests(provider):
@@ -261,3 +262,100 @@ class SocialAccountTests(TestCase):
             EmailAddress.objects.filter(user=user,
                                         email=user_email(user)).exists()
         )
+
+    @override_settings(
+        ACCOUNT_EMAIL_REQUIRED=True,
+        ACCOUNT_UNIQUE_EMAIL=True,
+        ACCOUNT_USERNAME_REQUIRED=True,
+        ACCOUNT_AUTHENTICATION_METHOD='email',
+        SOCIALACCOUNT_AUTO_SIGNUP=True)
+    def test_email_address_clash_username_required(self):
+        """Test clash on both username and email"""
+        request, resp = self._email_address_clash(
+            'test',
+            'test@test.com')
+        self.assertEqual(
+            resp['location'],
+            reverse('socialaccount_signup'))
+
+        # POST different username/email to social signup form
+        request.method = 'POST'
+        request.POST = {
+            'username': 'other',
+            'email': 'other@test.com'}
+        resp = signup(request)
+        self.assertEqual(
+            resp['location'], '/accounts/profile/')
+        user = get_user_model().objects.get(
+            **{account_settings.USER_MODEL_EMAIL_FIELD:
+               'other@test.com'})
+        self.assertEqual(user_username(user), 'other')
+
+    @override_settings(
+        ACCOUNT_EMAIL_REQUIRED=True,
+        ACCOUNT_UNIQUE_EMAIL=True,
+        ACCOUNT_USERNAME_REQUIRED=False,
+        ACCOUNT_AUTHENTICATION_METHOD='email',
+        SOCIALACCOUNT_AUTO_SIGNUP=True)
+    def test_email_address_clash_username_not_required(self):
+        """Test clash while username is not required"""
+        request, resp = self._email_address_clash(
+            'test',
+            'test@test.com')
+        self.assertEqual(
+            resp['location'],
+            reverse('socialaccount_signup'))
+
+        # POST email to social signup form (username not present)
+        request.method = 'POST'
+        request.POST = {
+            'email': 'other@test.com'}
+        resp = signup(request)
+        self.assertEqual(
+            resp['location'], '/accounts/profile/')
+        user = get_user_model().objects.get(
+            **{account_settings.USER_MODEL_EMAIL_FIELD:
+               'other@test.com'})
+        self.assertNotEqual(user_username(user), 'test')
+
+    @override_settings(
+        ACCOUNT_EMAIL_REQUIRED=True,
+        ACCOUNT_UNIQUE_EMAIL=True,
+        ACCOUNT_USERNAME_REQUIRED=False,
+        ACCOUNT_AUTHENTICATION_METHOD='email',
+        SOCIALACCOUNT_AUTO_SIGNUP=True)
+    def test_email_address_clash_username_auto_signup(self):
+        # Clash on username, but auto signup still works
+        request, resp = self._email_address_clash('test', 'other@test.com')
+        self.assertEqual(
+            resp['location'], '/accounts/profile/')
+        user = get_user_model().objects.get(
+            **{account_settings.USER_MODEL_EMAIL_FIELD:
+               'other@test.com'})
+        self.assertNotEqual(user_username(user), 'test')
+
+    def _email_address_clash(self, username, email):
+        User = get_user_model()
+        # Some existig user
+        exi_user = User()
+        user_username(exi_user, 'test')
+        user_email(exi_user, 'test@test.com')
+        exi_user.save()
+
+        # A social user being signed up...
+        account = SocialAccount(
+            provider='twitter',
+            uid='123')
+        user = User()
+        user_username(user, username)
+        user_email(user, email)
+        sociallogin = SocialLogin(user=user, account=account)
+
+        # Signing up, should pop up the social signup form
+        factory = RequestFactory()
+        request = factory.get('/accounts/twitter/login/callback/')
+        request.user = AnonymousUser()
+        SessionMiddleware().process_request(request)
+        MessageMiddleware().process_request(request)
+        resp = complete_social_login(request, sociallogin)
+        return request, resp